Crate rustc_codegen_spirv

Welcome to the API documentation for the rust-gpu project, this API is unstable and mainly intended for developing on the project itself. This is the API documentation for rustc_codegen_spirv which is not that useful on its own. You might also be interested in the following crates. There’s also the Rust GPU Dev Guide which contains more user-level information on how to use and setup rust-gpu.

spirv-builder

spirv-std

spirv-tools

spirv-tools-sys

Functions

__rustc_codegen_backend

This is the entrypoint for a hot plugged rustc_codegen_spirv